+++ rest_framework_services/dispatch/__init__.py
@@ -10,6 +10,9 @@ from rest_framework_services.dispatch.adispatch_spec import adispatch_spec
 from rest_framework_services.dispatch.aenforce_affordances import aenforce_affordances
 from rest_framework_services.dispatch.arender_for_audience import arender_for_audience
 from rest_framework_services.dispatch.arender_spec_output import arender_spec_output
+from rest_framework_services.dispatch.aunmet_operation_affordance import (
+    aunmet_operation_affordance,
+)
 from rest_framework_services.dispatch.base_pool import base_pool
 from rest_framework_services.dispatch.base_serializer_context import base_serializer_context
 from rest_framework_services.dispatch.build_offline_context import build_offline_context
@@ -34,12 +37,16 @@ from rest_framework_services.dispatch.renderable_serializer_class import (
     renderable_serializer_class,
 )
 from rest_framework_services.dispatch.unguarded_specs import unguarded_specs
+from rest_framework_services.dispatch.unmet_operation_affordance import (
+    unmet_operation_affordance,
+)
 
 __all__ = [
     "adispatch_spec",
     "aenforce_affordances",
     "arender_for_audience",
     "arender_spec_output",
+    "aunmet_operation_affordance",
     "base_pool",
     "base_serializer_context",
     "build_offline_context",
@@ -54,4 +61,5 @@ __all__ = [
     "render_spec_output",
     "renderable_serializer_class",
     "unguarded_specs",
+    "unmet_operation_affordance",
 ]